Dedupe DKVP-reader and NIDX-reader source files
